130 исправил все ошибки согласно замечаний подкорректировал стили
