+++ client/app/modules/CourseOptions/components/CourseOptionButton/CourseOptionButton.jsx
@@ -1,7 +1,7 @@
 import './CourseOptionButton.scss';
 import { MDBBtn, MDBPopover, MDBPopoverBody } from 'mdbreact';
 import React from 'react';
-import {Link} from 'react-router-dom'
+import {Link} from 'react-router-dom';
 const CourseOptionButton = (props) => {
 
     return (
